feat: Add a background job to handle article deletions
This came about as I was reading and experimentin with ActiveRecord
associations and how the `:destroy_async` works with the `dependent:` on
records. I tried it and it failed because I had database-level
validations for the foreign keys. So I decided to write a job that
handles deletion of articles so that it gives the user a faster response
time instead of them having to wait for the deletion to complete.

I simulated this by seeding an article with 1000 comments and deleting
it with and without the background job and I prefer it to be performed
in the background. It was a far superior user experience.