feat: Add a deleted field to the articles table
This update adds the `deleted` field to the articles table so that it's
easy to track which article has been deleted. This helps to make the
background job processing easier by first setting the deleted field to
`true`. This effectively causes it to not be in the list of articles
returned in the view.

Then the background job takes care of deleting the article itself.